added tests for Rect

+++ pc/src/sdp/group2/simulator/Visualizator.java
@@ -93,10 +93,10 @@ public class Visualizator extends JFrame {
     		for (int i = 1; i < N + 1; i++) {
     			Point p0 = pitchOutline.get((i - 1) % N);
     			Point p1 = pitchOutline.get(i % N);
-    			g.drawLine (Millimeter.mm2pix(p0.x), 
-    					Millimeter.mm2pix(p0.y),
-    					Millimeter.mm2pix(p1.x),
-    					Millimeter.mm2pix(p1.y));
+    			g.drawLine((int) Millimeter.mm2pix(p0.x),
+                        (int) Millimeter.mm2pix(p0.y),
+                        (int) Millimeter.mm2pix(p1.x),
+                        (int) Millimeter.mm2pix(p1.y));
     		}
     		
     		for(Zone zone : pitch.getAllZoneOutline()) {
@@ -106,10 +106,10 @@ public class Visualizator extends JFrame {
     			for (int i = 1; i < N + 1; i++) {
         			Point p0 = outline.get((i - 1) % N);
         			Point p1 = outline.get(i % N);
-        			g.drawLine (Millimeter.mm2pix(p0.x), 
-        					Millimeter.mm2pix(p0.y),
-        					Millimeter.mm2pix(p1.x),
-        					Millimeter.mm2pix(p1.y));
+        			g.drawLine((int) Millimeter.mm2pix(p0.x),
+                            (int) Millimeter.mm2pix(p0.y),
+                            (int) Millimeter.mm2pix(p1.x),
+                            (int) Millimeter.mm2pix(p1.y));
         		}
     			
     			g.setColor(new Color(0, 0, 100));
@@ -122,9 +122,9 @@ public class Visualizator extends JFrame {
     		g.setColor(new Color(255, 0, 0));
     		
     		Ball ball = pitch.getBall();
-    		int r = Millimeter.mm2pix(ball.getRadius());
-    		int x = Millimeter.mm2pix(ball.getPosition().x - r);
-    		int y = Millimeter.mm2pix(ball.getPosition().y - r);
+    		int r = (int) Millimeter.mm2pix(ball.getRadius());
+    		int x = (int) Millimeter.mm2pix(ball.getPosition().x - r);
+    		int y = (int) Millimeter.mm2pix(ball.getPosition().y - r);
     		g.fillOval(x, y, 2 * r, 2 * r);
      
     	}
